requirements.txt
asgiref==3.6.0
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
Cython==0.29.32
Django==4.1.4
django-adminlte3==0.1.6
django-compat==1.0.15
django-crispy-forms==1.14.0
django-filter==22.1
django-stdimage==6.0.1
djangorestframework==3.14.0
djhtml==1.5.2
gunicorn==20.1.0
idna==3.4
loguru==0.6.0
Markdown==3.4.1
Pillow==9.3.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pytz==2022.7
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
stripe==5.0.0
tomli==2.0.1
urllib3==1.26.13
whitenoise==6.3.0
